+++ src/main/java/br/com/fa7/firststepinagile/entities/Story.java
@@ -26,7 +26,7 @@ public class Story implements Serializable{
 	
 	private String description;
 	
-	private Integer priority;
+	private Double priority;
 	
 	private String color;
 	
@@ -71,11 +71,11 @@ public class Story implements Serializable{
 		this.description = description;
 	}
 
-	public Integer getPriority() {
+	public Double getPriority() {
 		return priority;
 	}
 
-	public void setPriority(Integer priority) {
+	public void setPriority(Double priority) {
 		this.priority = priority;
 	}
